Separate tests for the recipient
